Changed how legality of move is checked
Almost everything from game logic now in place: legality of a move will now be determined upon selection by the model. Need to implement and have begun coding a form of maze search to check if a wall placement violates the 'blocking constraint' of quoridor. After this game logic should be fully in place.